+++ pkg/utils/md5.go
@@ -12,6 +12,5 @@ import (
 func EncodeMD5(value string) string {
 	m := md5.New()
 	m.Write([]byte(value))
-
 	return hex.EncodeToString(m.Sum(nil))
 }