+++ scripts/migration/postgres/0015_password_use_bcrypt.down.sql
@@ -1,0 +1,12 @@
+ALTER TABLE p_user ADD COLUMN salt VARCHAR(32) NOT NULL DEFAULT '';
+
+UPDATE
+	p_user
+SET
+	salt = split_part(password, ':', -1),
+	password = split_part(password, ':', 1);
+
+ALTER TABLE p_user 
+ALTER COLUMN password TYPE VARCHAR(64),
+ALTER COLUMN password SET NOT NULL,
+ALTER COLUMN password SET DEFAULT '';