+++ scripts/migration/postgres/0015_password_use_bcrypt.up.sql
@@ -1,0 +1,11 @@
+ALTER TABLE p_user 
+ALTER COLUMN password TYPE VARCHAR(255) 
+ALTER COLUMN password SET NOT NULL
+ALTER COLUMN password SET DEFAULT '';
+
+UPDATE
+	p_user
+SET
+	password = concat_ws(':', password, salt);
+
+ALTER TABLE p_user DROP COLUMN salt;